Removed address_util.go file.
